backend/requirements.txt
Django==3.2.6
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
gunicorn==20.0.4
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2021.3
sqlparse==0.4.2
drf-extra-fields==3.4.0
